Start with a failing integration test & some refactorings
